backend/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6

# 데이터 처리
pandas==2.1.3
numpy>=1.26.0

# 날짜/시간 처리
python-dateutil==2.8.2

# 환경 변수
python-dotenv==1.0.0

# 데이터베이스 (MySQL)
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# HTTP 클라이언트
httpx==0.25.2


